Add URL validations and handling
When submitting links, the URL should be checked for conformance with
RFC2396 using the URI module. Only valid http and https URIs should be
allowed.

URIs without a protocol can also be allowed, but in order to properly
link to an external resource a default protocol (eg http://) needs to
prepended to it. Otherwise most browsers will treat the link as a
relative link.

To prevent unexpected surprises, users should be able to see the host a
link points to in its listing. Additionally, the header display in an
item listing should be consistent with the display in the index

*   Add URI validations
*   Extract host from urls
*   Normalize schemeless URLs
*   Normalize display of link title, host, and metadata
*   Display link icon in headers