Squashed commit of the following:
In order to support posts besides links to outside sites, we renamed the
Link model to Post, and added a boolean to specify whether a Post is a
link.

Posts can now have descriptions (which like comments are formatted in
Markdown)

Posts should be sorted by their total votes. TODO: Cache the total votes
so they can be pulled directly with a db query.

*   Rename links to posts
*   Add descriptions to posts
*   Add non-link posts
*   Sort posts by votes_total
*   Fix styling errors on mobile